+++ tests/analysis/test_cards.py
@@ -26,7 +26,7 @@ def test_series_card_data_sources(api_client, handle_usage_statistics):
     password = "secret"
     user = UserFactory(password=password)
     surface = SurfaceFactory(creator=user)
-    func1 = AnalysisFunction.objects.get(name="Test implementation")
+    func1 = AnalysisFunction.objects.get(name="topobank.testing.test")
 
     topo1 = Topography2DFactory(surface=surface)
 
@@ -107,7 +107,7 @@ def test_series_card_if_no_successful_topo_analysis(
     user = UserFactory(password=password)
     ContentType.objects.get_for_model(Topography)
     ContentType.objects.get_for_model(Surface)
-    func1 = AnalysisFunction.objects.get(name="Test implementation")
+    func1 = AnalysisFunction.objects.get(name="topobank.testing.test")
 
     surf = SurfaceFactory(creator=user)
     topo = Topography1DFactory(surface=surf)  # also generates the surface